requirements.txt
#requirements.txt
Django==1.5.12
markdown==2.3.1
django-admin-bootstrapped==1.6.5
django-taggit==0.11.2
